Connected updateQuestion() api call to backend. refs #15167 @0.5h
Note: I do not know if the way we deal with updated Choices in QuestionServiceImpl is currently correct. I.e. if a choice is deleted and not re-added to the question, is it left orphaned? Not a super critical bug because it does not affect functionality, but worth looking into...